HW_024 API testing architecture refactoring
